travis ci plan test sharding
