+++ plan/plan_test.go
@@ -200,42 +200,3 @@ func TestResolveTfLintComponent(test *testing.T) {
 		})
 	}
 }
-
-func Test_buildTravisCI(t *testing.T) {
-	a := assert.New(t)
-	{
-		c := &config.Config{
-			TravisCI: &config.TravisCI{
-				Enabled: false,
-			},
-		}
-		p := &Plan{}
-		tr := p.buildTravisCI(c)
-		a.NotNil(tr)
-		a.False(tr.Enabled)
-	}
-
-	// ----
-	{
-		id1 := int64(123456789)
-		c := &config.Config{
-			Accounts: map[string]config.Account{
-				"foo": config.Account{
-					AccountID: &id1,
-				},
-			},
-			TravisCI: &config.TravisCI{
-				Enabled:        true,
-				AWSIAMRoleName: "rollin",
-				HubAccountName: "hub",
-			},
-		}
-		p := &Plan{}
-		tr := p.buildTravisCI(c)
-		a.Len(tr.AWSProfiles, 1)
-		a.Equal(tr.AWSProfiles[0].Name, "foo")
-		a.Equal(tr.AWSProfiles[0].ID, id1)
-		a.Equal(tr.AWSProfiles[0].Role, "rollin")
-		a.Equal(tr.AWSProfiles[0].HubAccountName, "hub")
-	}
-}
